+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
 v0.71.0
 
 - Java, Node, Swift: changed IdentityKeyStore.saveIdentity to return an enum.
+
+- Node: RegistrationService.registerAccount takes account password as a string.